pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buyer-scout"
version = "0.1.0"
description = "Lead acquisition CLI using BBB Playwright crawling and SerpAPI discovery"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "playwright>=1.48.0",
  "requests>=2.32.0",
  "python-dateutil>=2.9.0",
  "python-dotenv>=1.0.1"
]

[project.scripts]
buyer-scout = "buyer_scout.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
